refactor: : 변경된 메인화면 UI 반영 & '헤더 - 촬영 섹션 - 이미지 불러오기' 영역으로 메인 화면 분리
